Fixed a logic comparison
Within the truth table for the denying of character groups to repeat consecutively function, the truth table would only allow this function to run if both upper and lowercase characters were used; this did not take into account the numbers are symbols. A simple and elegant fix was to implement a method for counting a boolean array, then ensuring that which is greater than one; thusly resulting in more than one group for this feature to run.